pictureVerify/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
get==0.0.0
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
post==0.0.0
public==0.0.0
query-string==0.0.0
setupfiles==0.0.50
SQLAlchemy==1.1.14
urllib3==1.22
Werkzeug==0.12.2
